+++ CHANGELOG.rst
@@ -9,6 +9,9 @@ Unreleased
 * feat(Instrument.holders_by_date): new method
 * feat(Instrument.client_type_history): new method
 * feat(Instrument.historic_data): new method
+* feat(Instrument.on_date): new method which returns an object with the following methods:
+
+  * ``price``
 
 v0.40.0
 -------